+++ src/main/java/com/example/plusproject/shoppingmall/entity/ShoppingMall.java
@@ -8,6 +8,7 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -15,6 +16,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import java.time.LocalDate;
 
 @Getter
+@Setter
 @Entity
 @NoArgsConstructor
 @Table(name = "shopping_malls_of_seoul")
@@ -24,24 +26,60 @@ public class ShoppingMall extends BaseEntity {
     private Long id;
 
     // 우선 17개 필드값만 입력해 두었습니다. validation 필요한 것 체크할 예정
+
+    // 상호
+    @Column(nullable = false)
     private String companyName;
+
+    // 쇼핑몰명
+    @Column(unique = true)
     private String storeName;
+
+    // 도메인명
     private String domainName;
+
+    // 전화번호
     private String phoneNumber;
+
+    // 운영자 이메일
     private String operatorEmail;
+
+    // 영업형태
     private String businessType;
+
+    // 최초신고일자
     private LocalDate registrationDate;
+
+    // 회사주소
     private String companyAddress;
+
+    // 업소상태
     @Enumerated(EnumType.STRING)
     private StoreStatus storeStatus;
+
+    // 전체평가
     @Enumerated(EnumType.STRING)
     private TotalRating totalRating;
+
+    // 주요취급품목
     private String mainProducts;
+
+    // 청약철회가능여부
     private String subscriptionWithdrawalAvailable;
+
+    // 초기화면필수항목
     private String homepageRequiredItems;
+
+    // 이용약관준수정도
     private String termsOfServiceCompliance;
+
+    // 배송예정일표시
     private String estimateDeliveryDateDisplay;
+
+    // 철회시배송비부담
     private String withdrawalShippingCostResponsibility;
+
+    // 모니터링날짜
     private LocalDate monitoringDate;
 
     @ManyToOne(fetch = FetchType.LAZY)
