--- src/main/java/com/example/plusproject/shoppingmall/openapi/StoreRepository.java
@@ -1,7 +1,0 @@
-package com.example.plusproject.shoppingmall.openapi;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoreRepository extends JpaRepository<Store, Long> {
-    boolean existsByStoreName(String storeName);
-}
